Generate allIdpNames; find idp type safely
