+++ hoauth2-demo/hoauth2-demo.cabal
@@ -32,7 +32,6 @@ executable hoauth2-demo
   autogen-modules:    Paths_hoauth2_demo
   other-modules:
     App
-    AppEnv
     Env
     Idp
     Paths_hoauth2_demo
@@ -48,6 +47,7 @@ executable hoauth2-demo
     DerivingStrategies
     ImportQualifiedPost
     InstanceSigs
+    LambdaCase
     OverloadedStrings
     PolyKinds
     RecordWildCards
